Removes the tmux navigation
The default mappings shades the completion mappings with neo-complete.
While there is a way to change the mappings of the plug-in these
mappings cannot be to complex. This is due to limitations on the tmux
side. Tmux only supports simple one key bindings so far. With this
commit I revert to the old behaviour until I find another solution.